Append a postfix to the icu target outputs on fuchsia (where target output is hardcoded)
On Fuchsia the target output names are hard-coded (instead of derived from
the name of the target) which will cause a GN error saying that two targets
have the same outputs.

So if we're on fuchsia, we'll append a "_hidden_visibility" postfix to
the target names to ensure they are unique.

This is a follow-up change to [0].

[0] https://chromium-review.googlesource.com/c/chromium/deps/icu/+/2236407

Change-Id: Ic3dc9347115598c9c007797941b3ba2c7436394c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/2248560
Reviewed-by: Frank Tang <ftang@chromium.org>