Set icu_disable_thin_archive on non-component chromeos builds
icu_disable_thin_archive is a GN arg which is set to true on
chrome-icu package of chromeos. See:
https://chromium.googlesource.com/chromiumos/overlays/chromiumos-overlay/+/refs/heads/master/chromeos-base/chrome-icu/chrome-icu-9999.ebuild#105

The other day I got an uprev failue due to a bug in a GN file.
See https://crbug.com/1129473 for the details. Perhaps it is
worthwhile to set a default value of this argument, so that
this type of error can be found much earlier, in a CQ step of
chromium.

Bug: chromium:1132766
Test: gn args pass with both is_component_build=true/false on chromeos
Change-Id: I27bcfa60da556e91afe2600a6cd3aadfddfedfe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/2436994
Reviewed-by: Honglin Yu <honglinyu@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Frank Tang <ftang@chromium.org>