Remove explicit std::atomic<NumberFormaterImpl*> template instantiation
The template cannot be explicitly instantiated here because doing so
results in errors while attempting to instantiate operator-=:

.../atomic(1864,80): error: arithmetic on a pointer to an incomplete type 'icu_68::number::impl::NumberRangeFormatterImpl'
      _Tp* operator-=(ptrdiff_t __op) _NOEXCEPT          {return fetch_sub(__op) - __op;}

I encountered this error while disabling the
exclude_from_explicit_instantiation macros that appear in libc++, and normally
appear on this operator-= implementation. I added this define to my build to do
it:
  -D_LIBCPP_HIDE_FROM_ABI=_LIBCPP_HIDDEN
In the future, I would like to disable the exclude_from_explicit_instantiation
attributes more cleanly, and at that point, this ICU instantiation will become
an issue.

I tested a Windows debug DLL build, and this instantiation does not
appear to be necessary.

R=ftang@chromium.org, thakis@chromium.org

Bug: 1090975
Change-Id: Icda96dbc32f77509e3a8801f0f54068a02e86718
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/2649173
Reviewed-by: Nico Weber <thakis@chromium.org>