Fix empty coll_tree resources issue
Needed to implement intl-enumeration API correctly

Remove the empty resource
coll/zh.res
zh{
  collations{
        big5han{}
        gb2312han{}
        unihan{}
  }
}

The order of the filter in the json file is tricky, if we place the line
below the "zh" filter, it will actually add an empty resource there.
So we have to swap the order of these two rules to remove them.
	"-/collations/*/Sequence"


Bug: v8:10743
Change-Id: Ifd3e984ab7e83163d2e0f0e7b2e119de563d60a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/2964068
Reviewed-by: Jungshik Shin <jshin@chromium.org>