[fuchsia] Limit the visibility of ICU targets
This narrows down the code that uses ICU in Fuchsia to a very specific
subdirectory of the build tree.

Fuchsia needs this because it requires very specific choices
in the ICU library, and we use visibility to force all
dependencies to take in the code that makes those
choices correctly.

Bug: fuchsia:115614
Change-Id: I731ad5fbbdee675cc2432e6578d6d1808a1b529d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/4035094
Reviewed-by: Frank Tang <ftang@chromium.org>