[fuchsia] Add a slot for extra compilation flags
Fuchsia sometimes needs these.

Fuchsia sometimes requires extra compilation flags for ICU to adapt it to
its current toolchain. Since it takes a while for ICU to roll through
Fuchsia, it can take a long time from an ICU commit to a fix rolling into
Fuchsia. This flag allows us to define the flag ahead of time in
//build/icu.gni, and remove the rollout issues.

Bug: fuchsia:121184
Change-Id: Iee30376581be789547175c6cf596e3357a3d2549
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/4258305
Reviewed-by: Frank Tang <ftang@chromium.org>