Add language names for Translate
The following 11 languages are used by Translate so need to be added
to the Accept-Language list:

  ay, bho, bm, doi, dv, ilo, lus, mai, mni, sa, ts

"as" and "kok" will also be added as Translate languages, but are already supported by ICU.

We only need the language name from ICU in all UI languages and the
native name of each of the languages.

The following files are added in source/data/lang to add the native
display name of {ay,dv,ilo,lus,ts} along with the patch file and
a note about them in README.chromium.

ay.txt
dv.txt
ilo.txt
lus.txt
ts.txt

The data size difference is as follows:
         Old      New      Delta
chromeos 11718496 11730784 12288
common   10560960 10573104 12144

Bug: 1313213
Change-Id: I219e68023fbadb03fcedbe0e8d02237cedbd73cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/4585451
Reviewed-by: Jungshik Shin <jshin@chromium.org>